+++ src/app.d.ts
@@ -16,6 +16,15 @@ declare global {
 		readonly effectiveType?: 'slow-2g' | '2g' | '3g' | '4g';
 		readonly downlink?: number;
 		readonly rtt?: number;
+		readonly type?:
+			| 'bluetooth'
+			| 'cellular'
+			| 'ethernet'
+			| 'none'
+			| 'wifi'
+			| 'wimax'
+			| 'other'
+			| 'unknown';
 	}
 	interface Navigator {
 		readonly connection?: NetworkInformation;